+++ combustion_ble/utilities/asyncio_utils.py
@@ -3,6 +3,8 @@
 import asyncio
 from typing import Any, Callable, Coroutine
 
+from bleak import BleakError
+
 from combustion_ble.logger import LOGGER
 
 DoneCallback = Callable[[asyncio.Task], None]
@@ -13,6 +15,19 @@ def _done_callback(task: asyncio.Task[Any]):
         task.result()
     except asyncio.CancelledError:
         pass  # Task cancellation should not be logged as an error.
+    except BleakError as be:
+        # This is raised specifically by the `corebluetooth` backend.
+        # It remains to be seen if this is a good idea or not. Many bleak operations are retried by us,
+        # and a single failure isn't something to be concerned about in most cases.
+        # TODO: handle other backend implementations.
+        if be.__str__() == "disconnected":
+            LOGGER.debug(
+                "Exception raised by task as a result of client disconnect = %r",
+                task,
+                exc_info=True,
+            )
+        else:
+            LOGGER.exception("Exception raised by task = %r", task)
     except Exception:
         LOGGER.exception("Exception raised by task = %r", task)
     if task.cancelled():