+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Allow BleManager to be stopped and restarted (#14)
 
 ## [v0.1.1](https://github.com/legrego/combustion_ble/releases/tag/v0.1.1) - 2024-01-10
 - Fix Documentation