+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 - Allow BleManager to be stopped and restarted (#14)
 - Add device listeners to respond to device changes (#15)
+- Prevent concurrent GATT read requests (#18)
 
 ## [v0.1.1](https://github.com/legrego/combustion_ble/releases/tag/v0.1.1) - 2024-01-10
 - Fix Documentation